# perf: improve memory usage during request lifecycle