--- a/server/src/client_api.rs
+++ b/server/src/client_api.rs
@@ -1,4 +1,5 @@
 use crate::error::{EdgeError, FeatureError};
+use crate::filters::{name_prefix_filter, project_filter, FeatureFilterSet};
 use crate::http::feature_refresher::FeatureRefresher;
 use crate::metrics::client_metrics::MetricsCache;
 use crate::tokens::cache_key;
@@ -65,8 +66,29 @@ async fn resolve_features(
         .get(&edge_token.token)
         .map(|e| e.value().clone())
         .ok_or(EdgeError::AuthorizationDenied)?;
+
+    let filters = filter_query.into_inner();
+    let query = unleash_types::client_features::Query {
+        tags: None,
+        projects: Some(validated_token.projects.clone()),
+        name_prefix: filters.name_prefix.clone(),
+        environment: validated_token.environment.clone(),
+        inline_segment_constraints: Some(false),
+    };
+
+    let filter_set = if let Some(name_prefix) = filters.name_prefix {
+        FeatureFilterSet::from(Box::new(name_prefix_filter(name_prefix)))
+    } else {
+        FeatureFilterSet::default()
+    }
+    .with_filter(project_filter(&edge_token));
+
     let client_features = match req.app_data::<Data<FeatureRefresher>>() {
-        Some(refresher) => refresher.features_for_token(validated_token.clone()).await,
+        Some(refresher) => {
+            refresher
+                .features_for_filter(validated_token.clone(), filter_set)
+                .await
+        }
         None => features_cache
             .get(&cache_key(&edge_token))
             .map(|features| features.value().clone())
@@ -77,32 +99,12 @@ async fn resolve_features(
                 ..client_features
             })
             .ok_or(EdgeError::ClientFeaturesFetchError(FeatureError::Retriable)),
-    };
-    let filters = filter_query.into_inner();
-    let query = unleash_types::client_features::Query {
-        tags: None,
-        projects: Some(validated_token.projects),
-        name_prefix: filters.name_prefix.clone(),
-        environment: validated_token.environment,
-        inline_segment_constraints: Some(false),
-    };
-    client_features
-        .map(|f| ClientFeatures {
-            query: Some(query),
-            features: f
-                .features
-                .into_iter()
-                .filter(|f| {
-                    filters
-                        .name_prefix
-                        .as_ref()
-                        .map(|prefix| f.name.starts_with(prefix))
-                        .unwrap_or(true)
-                })
-                .collect(),
-            ..f
-        })
-        .map(Json)
+    }?;
+
+    Ok(Json(ClientFeatures {
+        query: Some(query),
+        ..client_features
+    }))
 }
 #[utoipa::path(
     context_path = "/api/client",