--- a/server/src/lib.rs
+++ b/server/src/lib.rs
@@ -8,6 +8,7 @@ pub mod client_api;
 pub mod edge_api;
 #[cfg(not(tarpaulin_include))]
 pub mod error;
+pub mod filters;
 pub mod frontend_api;
 pub mod health_checker;
 pub mod http;