## What

This inverts the way we do filtering of the feature toggles when resolving data out of our cache. Previously we'd materialise the entire response, cloning each. Now we build the filtering at a higher layer and pass the filter in to the method that resolves the toggles. This means that the filter is applied **before** the cloning.

## Why

We're a little inefficient with the number of clones we apply when resolving a toggle. This means we stack allocate a **lot** of unnecessary memory. This reduces this to one clone (two on one cold path of retrieving toggles when they're not present in the cache). 

All in all, this doubles the response speed when working with a large number of toggles and halves the memory usage to achieve that.

## Benchmarks

### Baseline benchmark

Summary:
  Total:	14.3805 secs
  Slowest:	10.0794 secs
  Fastest:	0.9730 secs
  Average:	5.5344 secs
  Requests/sec:	7.8579
  

Response time histogram:
  0.973 [1]	|■
  1.884 [4]	|■■■■
  2.794 [7]	|■■■■■■■
  3.705 [7]	|■■■■■■■
  4.616 [9]	|■■■■■■■■■
  5.526 [18]	|■■■■■■■■■■■■■■■■■■
  6.437 [39]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  7.347 [10]	|■■■■■■■■■■
  8.258 [13]	|■■■■■■■■■■■■■
  9.169 [1]	|■
  10.079 [4]	|■■■■


Latency distribution:
  10% in 2.8589 secs
  25% in 4.7205 secs
  50% in 5.8351 secs
  75% in 6.4597 secs
  90% in 7.4248 secs
  95% in 8.9751 secs
  99% in 10.0794 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0015 secs, 0.9730 secs, 10.0794 secs
  DNS-lookup:	0.0004 secs, 0.0000 secs, 0.0016 secs
  req write:	0.0001 secs, 0.0000 secs, 0.0003 secs
  resp wait:	3.2266 secs, 0.0585 secs, 7.3207 secs
  resp read:	1.8054 secs, 0.0001 secs, 6.3262 secs

Status code distribution:
  [200]	107 responses
  [408]	6 responses

### Baseline memory usage:

After 1 prewarm up strike and benchmarks run: 4.0 GB

### Benchmark after patch
Summary:
  Total:	12.3002 secs
  Slowest:	6.2325 secs
  Fastest:	1.0501 secs
  Average:	2.9045 secs
  Requests/sec:	15.6909
  

Response time histogram:
  1.050 [1]	|■
  1.568 [1]	|■
  2.087 [5]	|■■■
  2.605 [65]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.123 [73]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.641 [30]	|■■■■■■■■■■■■■■■■
  4.160 [7]	|■■■■
  4.678 [5]	|■■■
  5.196 [3]	|■■
  5.714 [2]	|■
  6.232 [1]	|■


Latency distribution:
  10% in 2.3857 secs
  25% in 2.4603 secs
  50% in 2.8073 secs
  75% in 3.1863 secs
  90% in 3.5532 secs
  95% in 4.3718 secs
  99% in 6.2325 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0003 secs, 1.0501 secs, 6.2325 secs
  DNS-lookup:	0.0001 secs, 0.0000 secs, 0.0009 secs
  req write:	0.0000 secs, 0.0000 secs, 0.0004 secs
  resp wait:	1.7042 secs, 0.3664 secs, 3.4124 secs
  resp read:	1.1997 secs, 0.1047 secs, 3.1137 secs

Status code distribution:
  [200]	193 responses


### Memory usage after patch

After 1 prewarm up strike and benchmarks run: 2.2GB